take data from API

+++ app/src/main/java/com/example/test_task/Person.kt
@@ -1,3 +1,9 @@
 package com.example.test_task
 
-data class Person(val id: Long, val firstname: String, val lastname: String, val company: String, val img: String)
+data class Person(
+    val id: Int,
+    val firstname: String,
+    val lastname: String,
+    val company: String,
+    val img: String
+)
